VS round fix
git-svn-id: svn://svn.tuxfamily.org/svnroot/tesseract/main@1889 b412e182-2ed4-4078-95c4-27dc59f57fa9

+++ src/engine/command.cpp
@@ -4036,7 +4036,7 @@ ICOMMAND(floor, "f", (float *n), floatret(floor(*n)));
 ICOMMAND(ceil, "f", (float *n), floatret(ceil(*n)));
 ICOMMAND(round, "ff", (float *n, float *k),
 {
-    float step = *k;
+    double step = *k;
     double r = *n;
     if(step > 0)
     {
